LayoutTests/http/tests/security/contentSecurityPolicy/extensions/manifest-v2/img-src/manifest-v2-img-src-scheme-only-allowed-expected.txt

ALERT: PASS: Successfully loaded image.

Test that CSP for extensions allows scheme-only CSP for other directives.
